Docker/requirements.txt

streamlit
replicate
pdfminer.six
pytesseract
pdf2image
openai
requests
beautifulsoup4
tiktoken
